requirements.txt
pywebio==1.8.2
pywebio-battery==0.4.0
click==8.1.3
pywebview==4.1
PyYAML==6.0
pillow==9.5.0
aiofiles==23.1.0